Removed creation of userdata + Changed userdata to only the metatable + Added PushDouble and GetDouble + Created function GetUserDataFromTable + Added functions LoadLib for loading the utils scripts in the LuaWrapper + Renamed SetLuaBasePath to SetLuaBaseDirectory + Added SetLuaLibsDirectory + RegisterObjectFunctions take now in a ILuaMember pointer + Removed warrnings + Added a pointer to the ILuaMember in the metatable + Added utils script(enum) + LW_RegisterCObjectFunctions take now in a pointer to the object

+++ LuaGame/src/Game.cpp
@@ -27,7 +27,7 @@ int main()
 	float time = 0;
 	double accumulator = 0.0;
 
-	window.setFramerateLimit(FPS);
+	window.setFramerateLimit((unsigned int)FPS);
 	while (window.isOpen())
 	{
 		sf::Event event;
